Move App.js to components folder
