+++ src/components/App.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import { withStyles } from '@material-ui/core/styles';
-import Sidebar from './components/Sidebar';
-import ChatHeader from './components/ChatHeader';
-import Chat from './components/Chat';
+import Sidebar from './Sidebar';
+import ChatHeader from './ChatHeader';
+import Chat from './Chat';
 
-import { chats, messages } from './mock-data';
+import { chats, messages } from '../mock-data';
 
 const styles = theme => ({
   root: {
